Update all connected gates function receives array
Closes #24
